added more collectibles and updated game ui and level select

+++ HighRiseRunner/Assets/Scripts/CollectableRotate.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class CollectableRotate : MonoBehaviour
 {
-    [SerializeField] int rotateSpeed = 1; 
+    [SerializeField] float rotateSpeed = 1; 
 
     // Update is called once per frame
     void Update()
